Refactor code and unify js and css compress interface and object and add extra Type property
