+++ utils.go
@@ -3,6 +3,8 @@ package compress
 import (
 	"bytes"
 	"html/template"
+	"fmt"
+	"os"
 )
 
 func parseTmpl(t *template.Template, data map[string]string) (string, error) {
@@ -23,4 +25,21 @@ func (a argString) Get(i int, args ...string) (r string) {
 		r = args[0]
 	}
 	return
+}
+
+func logError(err string, args ...interface{}) {
+	err = fmt.Sprintf(err, args...)
+	fmt.Fprintln(os.Stderr, err)
+}
+
+func logInfo(info string, args ...interface{}) {
+	info = fmt.Sprintf(info, args...)
+	fmt.Fprintln(os.Stdout, info)
+}
+
+
+func errHtml(err string, args ...interface{}) string {
+	err = fmt.Sprintf("Beego Compress: "+err, args...)
+	fmt.Fprintln(os.Stderr, err)
+	return "<!-- " + err + " -->"
 }
